Split x-ms-version and user-agent from common to separate packages
