+++ BreakingChanges.md
@@ -7,7 +7,8 @@
 ## Version XX.XX.XX:
 
 ### All:
-- The library has been split into 4 different packages:
+- The library has been split into 5 different packages:
+    - azure-storage-common
     - azure-storage-blob
     - azure-storage-file
     - azure-storage-queue
@@ -15,7 +16,10 @@
 - The package `azure-storage` is now deprecated.
 - The classes that were directly under azure.storage, not under azure.storage.*(blob, file, queue, table), are now under azure.storage.common.
     - Example: azure.storage.retry becomes azure.storage.common.retry
-
+- The methods on the class SharedAccessSignature were moved into corresponding new subclasses in each service package.
+    - Example: the methods generate_blob and generate_container were moved to BlobSharedAccessSignature in the blob package. BlobSharedAccessSignature inherits generate_account from SharedAccessSignature in common.
+    - Please refer to the reference documentation for more info.
+    
 ## Version 0.34.0:
 
 ### All:
