+++ azure-storage-common/azure/storage/common/sharedaccesssignature.py
@@ -18,7 +18,7 @@ from ._common_conversion import (
     _sign_string,
     _to_str,
 )
-from ._constants import X_MS_VERSION
+from ._constants import DEFAULT_X_MS_VERSION
 from ._serialization import (
     url_quote,
     _to_utc_datetime,
@@ -27,455 +27,24 @@ from ._serialization import (
 
 class SharedAccessSignature(object):
     '''
-    Provides a factory for creating blob, queue, table, and file shares access 
-    signature tokens with a common account name and account key.  Users can either 
+    Provides a factory for creating account access
+    signature tokens with an account name and account key. Users can either
     use the factory or can construct the appropriate service and use the 
     generate_*_shared_access_signature method directly.
     '''
 
-    def __init__(self, account_name, account_key):
+    def __init__(self, account_name, account_key, x_ms_version=DEFAULT_X_MS_VERSION):
         '''
         :param str account_name:
             The storage account name used to generate the shared access signatures.
         :param str account_key:
-            The access key to genenerate the shares access signatures.
+            The access key to generate the shares access signatures.
+        :param str x_ms_version:
+            The service version used to generate the shared access signatures.
         '''
         self.account_name = account_name
         self.account_key = account_key
-
-    def generate_table(self, table_name, permission=None,
-                       expiry=None, start=None, id=None,
-                       ip=None, protocol=None,
-                       start_pk=None, start_rk=None,
-                       end_pk=None, end_rk=None):
-        '''
-        Generates a shared access signature for the table.
-        Use the returned signature with the sas_token parameter of TableService.
-
-        :param str table_name:
-            Name of table.
-        :param TablePermissions permission:
-            The permissions associated with the shared access signature. The 
-            user is restricted to operations allowed by the permissions. 
-            Required unless an id is given referencing a stored access policy 
-            which contains this field. This field must be omitted if it has been 
-            specified in an associated stored access policy.
-        :param expiry:
-            The time at which the shared access signature becomes invalid. 
-            Required unless an id is given referencing a stored access policy 
-            which contains this field. This field must be omitted if it has 
-            been specified in an associated stored access policy. Azure will always 
-            convert values to UTC. If a date is passed in without timezone info, it 
-            is assumed to be UTC.
-        :type expiry: datetime or str
-        :param start:
-            The time at which the shared access signature becomes valid. If 
-            omitted, start time for this call is assumed to be the time when the 
-            storage service receives the request. Azure will always convert values 
-            to UTC. If a date is passed in without timezone info, it is assumed to 
-            be UTC.
-        :type start: datetime or str
-        :param str id:
-            A unique value up to 64 characters in length that correlates to a 
-            stored access policy. To create a stored access policy, use 
-            set_blob_service_properties.
-        :param str ip:
-            Specifies an IP address or a range of IP addresses from which to accept requests.
-            If the IP address from which the request originates does not match the IP address
-            or address range specified on the SAS token, the request is not authenticated.
-            For example, specifying sip=168.1.5.65 or sip=168.1.5.60-168.1.5.70 on the SAS
-            restricts the request to those IP addresses.
-        :param str protocol:
-            Specifies the protocol permitted for a request made. The default value
-            is https,http. See :class:`~azure.storage.common.models.Protocol` for possible values.
-        :param str start_pk:
-            The minimum partition key accessible with this shared access 
-            signature. startpk must accompany startrk. Key values are inclusive. 
-            If omitted, there is no lower bound on the table entities that can 
-            be accessed.
-        :param str start_rk:
-            The minimum row key accessible with this shared access signature. 
-            startpk must accompany startrk. Key values are inclusive. If 
-            omitted, there is no lower bound on the table entities that can be 
-            accessed.
-        :param str end_pk:
-            The maximum partition key accessible with this shared access 
-            signature. endpk must accompany endrk. Key values are inclusive. If 
-            omitted, there is no upper bound on the table entities that can be 
-            accessed.
-        :param str end_rk:
-            The maximum row key accessible with this shared access signature. 
-            endpk must accompany endrk. Key values are inclusive. If omitted, 
-            there is no upper bound on the table entities that can be accessed.
-        '''
-        sas = _SharedAccessHelper()
-        sas.add_base(permission, expiry, start, ip, protocol)
-        sas.add_id(id)
-        sas.add_table_access_ranges(table_name, start_pk, start_rk, end_pk, end_rk)
-
-        # Table names must be signed lower case
-        resource_path = table_name.lower()
-        sas.add_resource_signature(self.account_name, self.account_key, 'table', resource_path)
-
-        return sas.get_token()
-
-    def generate_queue(self, queue_name, permission=None,
-                       expiry=None, start=None, id=None,
-                       ip=None, protocol=None):
-        '''
-        Generates a shared access signature for the queue.
-        Use the returned signature with the sas_token parameter of QueueService.
-
-        :param str queue_name:
-            Name of queue.
-        :param QueuePermissions permission:
-            The permissions associated with the shared access signature. The 
-            user is restricted to operations allowed by the permissions.
-            Permissions must be ordered read, add, update, process.
-            Required unless an id is given referencing a stored access policy 
-            which contains this field. This field must be omitted if it has been 
-            specified in an associated stored access policy.
-        :param expiry:
-            The time at which the shared access signature becomes invalid. 
-            Required unless an id is given referencing a stored access policy 
-            which contains this field. This field must be omitted if it has 
-            been specified in an associated stored access policy. Azure will always 
-            convert values to UTC. If a date is passed in without timezone info, it 
-            is assumed to be UTC.
-        :type expiry: datetime or str
-        :param start:
-            The time at which the shared access signature becomes valid. If 
-            omitted, start time for this call is assumed to be the time when the 
-            storage service receives the request. Azure will always convert values 
-            to UTC. If a date is passed in without timezone info, it is assumed to 
-            be UTC.
-        :type start: datetime or str
-        :param str id:
-            A unique value up to 64 characters in length that correlates to a 
-            stored access policy. To create a stored access policy, use 
-            set_blob_service_properties.
-        :param str ip:
-            Specifies an IP address or a range of IP addresses from which to accept requests.
-            If the IP address from which the request originates does not match the IP address
-            or address range specified on the SAS token, the request is not authenticated.
-            For example, specifying sip=168.1.5.65 or sip=168.1.5.60-168.1.5.70 on the SAS
-            restricts the request to those IP addresses.
-        :param str protocol:
-            Specifies the protocol permitted for a request made. The default value
-            is https,http. See :class:`~azure.storage.common.models.Protocol` for possible values.
-        '''
-        sas = _SharedAccessHelper()
-        sas.add_base(permission, expiry, start, ip, protocol)
-        sas.add_id(id)
-        sas.add_resource_signature(self.account_name, self.account_key, 'queue', queue_name)
-
-        return sas.get_token()
-
-    def generate_blob(self, container_name, blob_name, permission=None,
-                      expiry=None, start=None, id=None, ip=None, protocol=None,
-                      cache_control=None, content_disposition=None,
-                      content_encoding=None, content_language=None,
-                      content_type=None):
-        '''
-        Generates a shared access signature for the blob.
-        Use the returned signature with the sas_token parameter of any BlobService.
-
-        :param str container_name:
-            Name of container.
-        :param str blob_name:
-            Name of blob.
-        :param BlobPermissions permission:
-            The permissions associated with the shared access signature. The 
-            user is restricted to operations allowed by the permissions.
-            Permissions must be ordered read, write, delete, list.
-            Required unless an id is given referencing a stored access policy 
-            which contains this field. This field must be omitted if it has been 
-            specified in an associated stored access policy.
-        :param expiry:
-            The time at which the shared access signature becomes invalid. 
-            Required unless an id is given referencing a stored access policy 
-            which contains this field. This field must be omitted if it has 
-            been specified in an associated stored access policy. Azure will always 
-            convert values to UTC. If a date is passed in without timezone info, it 
-            is assumed to be UTC.
-        :type expiry: datetime or str
-        :param start:
-            The time at which the shared access signature becomes valid. If 
-            omitted, start time for this call is assumed to be the time when the 
-            storage service receives the request. Azure will always convert values 
-            to UTC. If a date is passed in without timezone info, it is assumed to 
-            be UTC.
-        :type start: datetime or str
-        :param str id:
-            A unique value up to 64 characters in length that correlates to a 
-            stored access policy. To create a stored access policy, use 
-            set_blob_service_properties.
-        :param str ip:
-            Specifies an IP address or a range of IP addresses from which to accept requests.
-            If the IP address from which the request originates does not match the IP address
-            or address range specified on the SAS token, the request is not authenticated.
-            For example, specifying sip=168.1.5.65 or sip=168.1.5.60-168.1.5.70 on the SAS
-            restricts the request to those IP addresses.
-        :param str protocol:
-            Specifies the protocol permitted for a request made. The default value
-            is https,http. See :class:`~azure.storage.common.models.Protocol` for possible values.
-        :param str cache_control:
-            Response header value for Cache-Control when resource is accessed
-            using this shared access signature.
-        :param str content_disposition:
-            Response header value for Content-Disposition when resource is accessed
-            using this shared access signature.
-        :param str content_encoding:
-            Response header value for Content-Encoding when resource is accessed
-            using this shared access signature.
-        :param str content_language:
-            Response header value for Content-Language when resource is accessed
-            using this shared access signature.
-        :param str content_type:
-            Response header value for Content-Type when resource is accessed
-            using this shared access signature.
-        '''
-        resource_path = container_name + '/' + blob_name
-
-        sas = _SharedAccessHelper()
-        sas.add_base(permission, expiry, start, ip, protocol)
-        sas.add_id(id)
-        sas.add_resource('b')
-        sas.add_override_response_headers(cache_control, content_disposition,
-                                          content_encoding, content_language,
-                                          content_type)
-        sas.add_resource_signature(self.account_name, self.account_key, 'blob', resource_path)
-
-        return sas.get_token()
-
-    def generate_container(self, container_name, permission=None, expiry=None,
-                           start=None, id=None, ip=None, protocol=None,
-                           cache_control=None, content_disposition=None,
-                           content_encoding=None, content_language=None,
-                           content_type=None):
-        '''
-        Generates a shared access signature for the container.
-        Use the returned signature with the sas_token parameter of any BlobService.
-
-        :param str container_name:
-            Name of container.
-        :param ContainerPermissions permission:
-            The permissions associated with the shared access signature. The 
-            user is restricted to operations allowed by the permissions.
-            Permissions must be ordered read, write, delete, list.
-            Required unless an id is given referencing a stored access policy 
-            which contains this field. This field must be omitted if it has been 
-            specified in an associated stored access policy.
-        :param expiry:
-            The time at which the shared access signature becomes invalid. 
-            Required unless an id is given referencing a stored access policy 
-            which contains this field. This field must be omitted if it has 
-            been specified in an associated stored access policy. Azure will always 
-            convert values to UTC. If a date is passed in without timezone info, it 
-            is assumed to be UTC.
-        :type expiry: datetime or str
-        :param start:
-            The time at which the shared access signature becomes valid. If 
-            omitted, start time for this call is assumed to be the time when the 
-            storage service receives the request. Azure will always convert values 
-            to UTC. If a date is passed in without timezone info, it is assumed to 
-            be UTC.
-        :type start: datetime or str
-        :param str id:
-            A unique value up to 64 characters in length that correlates to a 
-            stored access policy. To create a stored access policy, use 
-            set_blob_service_properties.
-        :param str ip:
-            Specifies an IP address or a range of IP addresses from which to accept requests.
-            If the IP address from which the request originates does not match the IP address
-            or address range specified on the SAS token, the request is not authenticated.
-            For example, specifying sip=168.1.5.65 or sip=168.1.5.60-168.1.5.70 on the SAS
-            restricts the request to those IP addresses.
-        :param str protocol:
-            Specifies the protocol permitted for a request made. The default value
-            is https,http. See :class:`~azure.storage.common.models.Protocol` for possible values.
-        :param str cache_control:
-            Response header value for Cache-Control when resource is accessed
-            using this shared access signature.
-        :param str content_disposition:
-            Response header value for Content-Disposition when resource is accessed
-            using this shared access signature.
-        :param str content_encoding:
-            Response header value for Content-Encoding when resource is accessed
-            using this shared access signature.
-        :param str content_language:
-            Response header value for Content-Language when resource is accessed
-            using this shared access signature.
-        :param str content_type:
-            Response header value for Content-Type when resource is accessed
-            using this shared access signature.
-        '''
-        sas = _SharedAccessHelper()
-        sas.add_base(permission, expiry, start, ip, protocol)
-        sas.add_id(id)
-        sas.add_resource('c')
-        sas.add_override_response_headers(cache_control, content_disposition,
-                                          content_encoding, content_language,
-                                          content_type)
-        sas.add_resource_signature(self.account_name, self.account_key, 'blob', container_name)
-
-        return sas.get_token()
-
-    def generate_file(self, share_name, directory_name=None, file_name=None,
-                      permission=None, expiry=None, start=None, id=None,
-                      ip=None, protocol=None, cache_control=None,
-                      content_disposition=None, content_encoding=None,
-                      content_language=None, content_type=None):
-        '''
-        Generates a shared access signature for the file.
-        Use the returned signature with the sas_token parameter of FileService.
-
-        :param str share_name:
-            Name of share.
-        :param str directory_name:
-            Name of directory. SAS tokens cannot be created for directories, so 
-            this parameter should only be present if file_name is provided.
-        :param str file_name:
-            Name of file.
-        :param FilePermissions permission:
-            The permissions associated with the shared access signature. The 
-            user is restricted to operations allowed by the permissions.
-            Permissions must be ordered read, create, write, delete, list.
-            Required unless an id is given referencing a stored access policy 
-            which contains this field. This field must be omitted if it has been 
-            specified in an associated stored access policy.
-        :param expiry:
-            The time at which the shared access signature becomes invalid. 
-            Required unless an id is given referencing a stored access policy 
-            which contains this field. This field must be omitted if it has 
-            been specified in an associated stored access policy. Azure will always 
-            convert values to UTC. If a date is passed in without timezone info, it 
-            is assumed to be UTC.
-        :type expiry: datetime or str
-        :param start:
-            The time at which the shared access signature becomes valid. If 
-            omitted, start time for this call is assumed to be the time when the 
-            storage service receives the request. Azure will always convert values 
-            to UTC. If a date is passed in without timezone info, it is assumed to 
-            be UTC.
-        :type start: datetime or str
-        :param str id:
-            A unique value up to 64 characters in length that correlates to a 
-            stored access policy. To create a stored access policy, use 
-            set_file_service_properties.
-        :param str ip:
-            Specifies an IP address or a range of IP addresses from which to accept requests.
-            If the IP address from which the request originates does not match the IP address
-            or address range specified on the SAS token, the request is not authenticated.
-            For example, specifying sip=168.1.5.65 or sip=168.1.5.60-168.1.5.70 on the SAS
-            restricts the request to those IP addresses.
-        :param str protocol:
-            Specifies the protocol permitted for a request made. The default value
-            is https,http. See :class:`~azure.storage.common.models.Protocol` for possible values.
-        :param str cache_control:
-            Response header value for Cache-Control when resource is accessed
-            using this shared access signature.
-        :param str content_disposition:
-            Response header value for Content-Disposition when resource is accessed
-            using this shared access signature.
-        :param str content_encoding:
-            Response header value for Content-Encoding when resource is accessed
-            using this shared access signature.
-        :param str content_language:
-            Response header value for Content-Language when resource is accessed
-            using this shared access signature.
-        :param str content_type:
-            Response header value for Content-Type when resource is accessed
-            using this shared access signature.
-        '''
-        resource_path = share_name
-        if directory_name is not None:
-            resource_path += '/' + _to_str(directory_name)
-        resource_path += '/' + _to_str(file_name)
-
-        sas = _SharedAccessHelper()
-        sas.add_base(permission, expiry, start, ip, protocol)
-        sas.add_id(id)
-        sas.add_resource('f')
-        sas.add_override_response_headers(cache_control, content_disposition,
-                                          content_encoding, content_language,
-                                          content_type)
-        sas.add_resource_signature(self.account_name, self.account_key, 'file', resource_path)
-
-        return sas.get_token()
-
-    def generate_share(self, share_name, permission=None, expiry=None,
-                       start=None, id=None, ip=None, protocol=None,
-                       cache_control=None, content_disposition=None,
-                       content_encoding=None, content_language=None,
-                       content_type=None):
-        '''
-        Generates a shared access signature for the share.
-        Use the returned signature with the sas_token parameter of FileService.
-
-        :param str share_name:
-            Name of share.
-        :param SharePermissions permission:
-            The permissions associated with the shared access signature. The 
-            user is restricted to operations allowed by the permissions.
-            Permissions must be ordered read, create, write, delete, list.
-            Required unless an id is given referencing a stored access policy 
-            which contains this field. This field must be omitted if it has been 
-            specified in an associated stored access policy.
-        :param expiry:
-            The time at which the shared access signature becomes invalid. 
-            Required unless an id is given referencing a stored access policy 
-            which contains this field. This field must be omitted if it has 
-            been specified in an associated stored access policy. Azure will always 
-            convert values to UTC. If a date is passed in without timezone info, it 
-            is assumed to be UTC.
-        :type expiry: datetime or str
-        :param start:
-            The time at which the shared access signature becomes valid. If 
-            omitted, start time for this call is assumed to be the time when the 
-            storage service receives the request. Azure will always convert values 
-            to UTC. If a date is passed in without timezone info, it is assumed to 
-            be UTC.
-        :type start: datetime or str
-        :param str id:
-            A unique value up to 64 characters in length that correlates to a 
-            stored access policy. To create a stored access policy, use 
-            set_file_service_properties.
-        :param str ip:
-            Specifies an IP address or a range of IP addresses from which to accept requests.
-            If the IP address from which the request originates does not match the IP address
-            or address range specified on the SAS token, the request is not authenticated.
-            For example, specifying sip=168.1.5.65 or sip=168.1.5.60-168.1.5.70 on the SAS
-            restricts the request to those IP addresses.
-        :param str protocol:
-            Specifies the protocol permitted for a request made. The default value
-            is https,http. See :class:`~azure.storage.common.models.Protocol` for possible values.
-        :param str cache_control:
-            Response header value for Cache-Control when resource is accessed
-            using this shared access signature.
-        :param str content_disposition:
-            Response header value for Content-Disposition when resource is accessed
-            using this shared access signature.
-        :param str content_encoding:
-            Response header value for Content-Encoding when resource is accessed
-            using this shared access signature.
-        :param str content_language:
-            Response header value for Content-Language when resource is accessed
-            using this shared access signature.
-        :param str content_type:
-            Response header value for Content-Type when resource is accessed
-            using this shared access signature.
-        '''
-        sas = _SharedAccessHelper()
-        sas.add_base(permission, expiry, start, ip, protocol)
-        sas.add_id(id)
-        sas.add_resource('s')
-        sas.add_override_response_headers(cache_control, content_disposition,
-                                          content_encoding, content_language,
-                                          content_type)
-        sas.add_resource_signature(self.account_name, self.account_key, 'file', share_name)
-
-        return sas.get_token()
+        self.x_ms_version = x_ms_version
 
     def generate_account(self, services, resource_types, permission, expiry, start=None,
                          ip=None, protocol=None):
@@ -524,7 +93,7 @@ class SharedAccessSignature(object):
             is https,http. See :class:`~azure.storage.common.models.Protocol` for possible values.
         '''
         sas = _SharedAccessHelper()
-        sas.add_base(permission, expiry, start, ip, protocol)
+        sas.add_base(permission, expiry, start, ip, protocol, self.x_ms_version)
         sas.add_account(services, resource_types)
         sas.add_account_signature(self.account_name, self.account_key)
 
@@ -563,7 +132,7 @@ class _SharedAccessHelper(object):
         if val:
             self.query_dict[name] = _to_str(val)
 
-    def add_base(self, permission, expiry, start, ip, protocol):
+    def add_base(self, permission, expiry, start, ip, protocol, x_ms_version):
         if isinstance(start, date):
             start = _to_utc_datetime(start)
 
@@ -575,7 +144,7 @@ class _SharedAccessHelper(object):
         self._add_query(_QueryStringConstants.SIGNED_PERMISSION, permission)
         self._add_query(_QueryStringConstants.SIGNED_IP, ip)
         self._add_query(_QueryStringConstants.SIGNED_PROTOCOL, protocol)
-        self._add_query(_QueryStringConstants.SIGNED_VERSION, X_MS_VERSION)
+        self._add_query(_QueryStringConstants.SIGNED_VERSION, x_ms_version)
 
     def add_resource(self, resource):
         self._add_query(_QueryStringConstants.SIGNED_RESOURCE, resource)
